Add Mantine library packages
Mantine is a set of multiple-use-case things library, we choose this library since it offers a lot of things and also the performance is great.

In the future, other features that are available in Mantine will be used first rather than using another library, like when we will be using markdown editor, we would use Mantine rich text editor instead of another library.