Migrating to react-beautiful-dnd
After using dnd-kit for a while, I noticed that the package has a significant bug that haven't got any resolve steps, such as jumping cards when card is dragged frequently.

I have search anywhere for this issue but still no resolve steps within the dnd-kit developers itself, so as for that I moved on to use react-beautiful-dnd since it was heavily used for similar things like Trello which was maintained by Atlassian.