Implement own Drag and Drop kit
While trying other libraries, it turns out that most of it has flaw that can't be usable in this situation, like no nested container scroll, wrong placeholder spacing, worst performance, etc.

So to top all of that, I decided to just implement our internal library for this, I was deciding to make this open source in the future. Still not the best, but works great 👍