Update collision strategy
In this new update, I decided to create and use closest corner strategy since it just makes more sense and gives more great result.

Also I add a collision visualizer to visualize the collisions with lines and numbers that are drawn on HTML canvas.

And I also add Auto sortable strategy for... idk?