Rework context menu
Instead of using external library, I decided to create a custom one instead, so that I can style the menu in the way I want.

It also uses Provider instead of custom event so that it can handle the shared data more properly in React way.

This will be published into another package in the future.