Switching to app directory
A lot of thought was taken into making this decision, one of which is the rendering layout that belongs to the NextJS 13 App Directory which makes it possible to share the layout with the subpages below it without re-fetching the whole/re-rendering.

And also the app directory has begun to be stable to use and also makes it easier for the metadata section.