Create utils to check permission for user and organization
While it was doing so much query, but I think, it's not a very big problem since if we look into Prisma include query, it's actually split the query into multiple query instead of left joining the previous query.

After all, we shouldn't expect that PostgreSQL doesn't have this much capacity to perform such a large query.

In the future we might want to find a better way for this, but for now, I think this was the better way to do this