Split refs for each menu type
On previous version, since it only use 1 ref to be shared to multiple different menu variation, when the menu changed (etc. opening sub menu), it doesn't update the ref with the latest variant, resulting on when we are clicking inside that new menu, the menu will be closed, since it was using old variant to check if the click was inside that variant or not.